pyproject.toml
[project]
name = "og_preview"
version = "2.0.0"
description = "Open Graph preview generator for articles on my personal website"
authors = [
    {name = "Andrew Sonin",email = "me@andrewsonin.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "notebook (>=7.4.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipython = "^9.2.0"
jupyter = "^1.1.1"

